git-svn-id: file:///home/nick/svnrepo/repo/indigo/trunk@1292 20aa11e4-6e6f-594c-a032-6291ace9ee65

+++ utils/StreamUtils.h
@@ -72,7 +72,7 @@ namespace StreamUtils
 	template <class T>
 	inline static void writeBasicVector(std::ostream& stream, const std::vector<T>& t)
 	{
-		const unsigned int sz = t.size();
+		const size_t sz = t.size();
 		writeBasic(stream, sz);
 		//for(size_t i=0; i<t.size(); ++i)
 		//	writeBasic(stream, t[i]);
